+++ src/digitalIdentity-frontend/src/app/pages/proof-templ/create-proof-template/services/pt-builder.service.spec.ts
@@ -1,0 +1,16 @@
+import { TestBed } from '@angular/core/testing';
+
+import { PtBuilderService } from './pt-builder.service';
+
+describe('PtBuilderService', () => {
+  let service: PtBuilderService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(PtBuilderService);
+  });
+
+  it('should be created', () => {
+    expect(service).toBeTruthy();
+  });
+});
